Python_scripts/requirements.txt
pySerial
PySide6
PyQt6
pyqtgraph
